chaostooling-demo/order-service/requirements.txt
flask==3.0.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0

